backend/requirements.txt
asgiref==3.7.2
atomicwrites==1.4.0
attrs==21.4.0
certifi==2021.10.8
cffi==1.16.0
charset-normalizer==2.0.11
colorama==0.4.4
cryptography==41.0.4
defusedxml==0.8.0rc2
Django==3.2.3
django-filter==21.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
flake8==4.0.1
idna==3.3
iniconfig==1.1.1
install==1.3.5
mccabe==0.6.1
oauthlib==3.2.2
packaging==21.3
Pillow==9.0.0
pluggy==0.13.1
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.1.0
pyparsing==3.0.7
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2021.3
regex==2023.8.8
requests==2.26.0
requests-oauthlib==1.3.1
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.8.0
urllib3==1.26.8
webcolors==1.13
